requirements.txt
# streamlit
streamlit
pillow

pytorch==1.8.0
torchvision==0.9.0
torchaudio==0.8.0
cudatoolkit=10.2

insightface==0.2.1
onnxruntime
numpy